SCRUM-51 - config: Refactor file structure
Change to one hydrosens folder file structure
