--- data/CN_lookup.csv
@@ -1,12 +1,0 @@
-0,1,2,3,4
-11,45,66,77,83
-12,36,60,73,79
-13,25,55,70,77
-21,57,73,82,86
-22,43,65,76,82
-23,32,58,72,79
-31,68,79,86,89
-32,49,69,79,84
-33,39,61,74,80
-41,69,84,88,91
-0,77,86,91,94
